1. 기본 자료형을 모두 래퍼 클래스로 변경 2. 면적의 자료형을 Double로 변경 3. CSV 파서 일차 구현(리스트, 데이터베이스 업로드 미구현)

+++ src/controller/MainNavigatorController.java
@@ -69,7 +69,7 @@ public class MainNavigatorController {
 					String code = arr[i].getCode();
 					String capital = arr[i].getCapital();
 					String location = arr[i].getLocation();
-					String area = Integer.toString(arr[i].getArea());
+					String area = Double.toString(arr[i].getArea());
 
 					table.setValueAt(name, i, 0);
 					table.setValueAt(code, i, 1);
